requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
astroid==2.3.1
Authlib==0.14.1
cffi==1.14.0
click==7.1.1
colorama==0.4.1
cryptography==2.9.2
DateTime==4.3
distlib==0.3.0
dnspython==1.16.0
filelock==3.0.12
Flask==1.1.2
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
numpy==1.17.3
passlib==1.7.2
protobuf==3.6.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.2
pytz==2020.1
six==1.12.0
typed-ast==1.4.0
uuid==1.30
virtualenv==20.0.5
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.0.0
zope.interface==5.1.0
